(core+) merging a ton of experimental branches
(core) re-order fix
Draw- order values cannot be negative, yet have
been permitted to be specified relative to a parent,
effectively meaning that a child cannot be drawn before
the object it is linked to. This change adds the option
to specify negative orders in relation to the parent.
Internally, the 0 <= n < 65536 constraint is still maintained.

(platform) non-buffer passing path regression

Refactoring the semantics for how shmif- accelerated
descriptor passing should propagate error messages
left the software- fallback broken. Partially fixed,
i.e. arcan_lwa works though there are still issues
with libretro frameserver.

(support) mouse tweaks to fit clock changes

also fixes the issue where several improper
events were being fired during a drag.

note: should really add a calibration surface
that times patterns for click, double click
etc.

(libretro) fixed regression in 3D core readback

the switch to agp for accelerated buffer passing
introduced some other issues to the default readback
approach, we still have issues with some core crashing
however.

(core) accelerated cursor used improper blendstate
(core) added support for tracking subsegment parent
(core) target_updatehandler rewr pending events
(core) frameserver first pbo transfer fix
(core) fixed subsegment getting wrong ffunc

+++ src/engine/arcan_video.h
@@ -680,7 +680,7 @@ arcan_errc arcan_video_objecttexmode(arcan_vobj_id id,
 	enum arcan_vtex_mode modes, enum arcan_vtex_mode modet);
 arcan_errc arcan_video_objectfilter(arcan_vobj_id id,
 	enum arcan_vfilter_mode filter);
-arcan_errc arcan_video_setzv(arcan_vobj_id id,unsigned short newzv);
+arcan_errc arcan_video_setzv(arcan_vobj_id id, int newzv);
 
 /* Object state retrieval */
 unsigned short arcan_video_getzv(arcan_vobj_id id);
